requirements.txt
dagster==1.8.5
dagster-webserver==1.8.5
psycopg2-binary==2.9.9
requests
